requirements.txt
langchain
python-dotenv
pyyaml
numpy
tqdm
llama-cpp-python
modelscope
fastapi
uvicorn
pydantic
pydantic-settings
pytest
